// packages/react-ii/src/types/country.ts
type Maybe<K extends string, V = object> = Partial<Record<K, V> | null> | null

export interface Country {
	name: {
		common: string
		official: string
		nativeName?: Maybe<NativeNames, { official: string; common: string }>
	}
	tld?: string[] | null
	cca2: string
	ccn3?: string | null
	cca3: string
	cioc?: string | null
	independent?: boolean | null
	status: string
	unMember: boolean
	currencies?: Maybe<CurrencyNames, { name: string; symbol: string }>
	idd: { root?: string | null; suffixes?: string[] | null }
	capital?: string[] | null
	altSpellings?: string[] | null
	region: string
	subregion?: string | null
	languages?: Maybe<Lang, string>
	translations: Maybe<Transl, { official: string; common: string }>
	latlng?: number[] | null
	landlocked: boolean
	borders?: string[] | null
	area: number
	demonyms?: Maybe<DemonymNames, { f: string; m: string }>
	flag: string
	maps: { googleMaps: string; openStreetMaps: string }
	population: number
	fifa?: string | null
	car: { signs?: string[] | null; side: string }
	timezones?: string[] | null
	continents?: string[] | null
	flags: { png: string; svg: string; alt?: string | null }
	coatOfArms: { png?: string | null; svg?: string | null }
	startOfWeek: string
	capitalInfo: { latlng?: number[] | null }
	postalCode?: { format: string; regex?: string | null } | null
	gini?: { [key: number]: number | null } | null
}

export type NativeNames =
	| 'ara'
	| 'eng'
	| 'fra'
	| 'fas'
	| 'ukr'
	| 'ind'
	| 'msa'
	| 'ces'
	| 'slk'
	| 'srp'
	| 'hmo'
	| 'tpi'
	| 'nau'
	| 'spa'
	| 'ron'
	| 'tvl'
	| 'zho'
	| 'sin'
	| 'tam'
	| 'afr'
	| 'nbl'
	| 'nso'
	| 'sot'
	| 'ssw'
	| 'tsn'
	| 'tso'
	| 'ven'
	| 'xho'
	| 'zul'
	| 'nfr'
	| 'swa'
	| 'dan'
	| 'fao'
	| 'smo'
	| 'tkl'
	| 'hin'
	| 'pih'
	| 'nep'
	| 'rus'
	| 'uzb'
	| 'fij'
	| 'hif'
	| 'mfe'
	| 'cnr'
	| 'gil'
	| 'por'
	| 'fil'
	| 'nor'
	| 'nld'
	| 'mri'
	| 'nzs'
	| 'lav'
	| 'tha'
	| 'arc'
	| 'ckb'
	| 'kin'
	| 'grn'
	| 'fin'
	| 'swe'
	| 'ita'
	| 'zdj'
	| 'mlt'
	| 'ell'
	| 'deu'
	| 'her'
	| 'hgm'
	| 'kwn'
	| 'loz'
	| 'ndo'
	| 'est'
	| 'ton'
	| 'urd'
	| 'glv'
	| 'aym'
	| 'que'
	| 'tur'
	| 'kir'
	| 'kon'
	| 'lin'
	| 'lat'
	| 'khm'
	| 'ber'
	| 'sqi'
	| 'hrv'
	| 'hye'
	| 'pau'
	| 'cat'
	| 'bjz'
	| 'hat'
	| 'sag'
	| 'bar'
	| 'kal'
	| 'tgk'
	| 'som'
	| 'div'
	| 'crs'
	| 'pap'
	| 'mlg'
	| 'bwg'
	| 'kck'
	| 'khi'
	| 'ndc'
	| 'nde'
	| 'nya'
	| 'sna'
	| 'toi'
	| 'zib'
	| 'kor'
	| 'tir'
	| 'jam'
	| 'mon'
	| 'ltz'
	| 'mkd'
	| 'dzo'
	| 'kaz'
	| 'gsw'
	| 'roh'
	| 'cha'
	| 'nno'
	| 'nob'
	| 'smi'
	| 'bos'
	| 'prs'
	| 'pus'
	| 'tuk'
	| 'slv'
	| 'isl'
	| 'bul'
	| 'kat'
	| 'gle'
	| 'run'
	| 'cal'
	| 'lit'
	| 'nrf'
	| 'ben'
	| 'mey'
	| 'vie'
	| 'hun'
	| 'amh'
	| 'bis'
	| 'mah'
	| 'jpn'
	| 'aze'
	| 'heb'
	| 'rar'
	| 'tet'
	| 'pol'
	| 'bel'
	| 'pov'
	| 'lao'
	| 'niu'
	| 'lua'
	| 'mya'

export type CurrencyNames =
	| 'SAR'
	| 'XAF'
	| 'IRR'
	| 'AED'
	| 'UAH'
	| 'XOF'
	| 'XCD'
	| 'USD'
	| 'IDR'
	| 'BND'
	| 'SGD'
	| 'CZK'
	| 'RSD'
	| 'PGK'
	| 'AUD'
	| 'BMD'
	| 'VES'
	| 'EUR'
	| 'RON'
	| 'TVD'
	| 'EGP'
	| 'ILS'
	| 'JOD'
	| 'CNY'
	| 'LKR'
	| 'ZAR'
	| 'GBP'
	| 'GGP'
	| 'TZS'
	| 'DKK'
	| 'FOK'
	| 'NZD'
	| 'MYR'
	| 'INR'
	| 'GIP'
	| 'SDG'
	| 'XPF'
	| 'NPR'
	| 'UZS'
	| 'FJD'
	| 'MUR'
	| 'KID'
	| 'PHP'
	| 'NOK'
	| 'MZN'
	| 'SRD'
	| 'SYP'
	| 'LSL'
	| 'CVE'
	| 'MRU'
	| 'SLL'
	| 'THB'
	| 'IQD'
	| 'NGN'
	| 'RWF'
	| 'LYD'
	| 'PYG'
	| 'BSD'
	| 'SHP'
	| 'SBD'
	| 'KMF'
	| 'GHS'
	| 'SEK'
	| 'BRL'
	| 'NAD'
	| 'TOP'
	| 'PKR'
	| 'IMP'
	| 'BOB'
	| 'KGS'
	| 'HNL'
	| 'HKD'
	| 'CUC'
	| 'CUP'
	| 'OMR'
	| 'BHD'
	| 'UYU'
	| 'YER'
	| 'KHR'
	| 'MAD'
	| 'BWP'
	| 'ZMW'
	| 'AMD'
	| 'TTD'
	| 'DZD'
	| 'BZD'
	| 'TWD'
	| 'TND'
	| 'HTG'
	| 'ANG'
	| 'CHF'
	| 'DOP'
	| 'CAD'
	| 'AOA'
	| 'TJS'
	| 'SZL'
	| 'SOS'
	| 'GYD'
	| 'MVR'
	| 'SCR'
	| 'AWG'
	| 'MGA'
	| 'ZWL'
	| 'KRW'
	| 'ERN'
	| 'JMD'
	| 'UGX'
	| 'KYD'
	| 'MNT'
	| 'GNF'
	| 'MKD'
	| 'BTN'
	| 'PEN'
	| 'KZT'
	| 'LRD'
	| 'QAR'
	| 'ARS'
	| 'FKP'
	| 'PAB'
	| 'AFN'
	| 'ISK'
	| 'BGN'
	| 'GEL'
	| 'BIF'
	| 'ALL'
	| 'COP'
	| 'JEP'
	| 'KWD'
	| 'BDT'
	| 'GMD'
	| 'VND'
	| 'HUF'
	| 'ETB'
	| 'VUV'
	| 'KPW'
	| 'GTQ'
	| 'RUB'
	| 'MXN'
	| 'CLP'
	| 'MOP'
	| 'CRC'
	| 'DJF'
	| 'JPY'
	| 'AZN'
	| 'MWK'
	| 'KES'
	| 'CKD'
	| 'SSP'
	| 'PLN'
	| 'BBD'
	| 'TRY'
	| 'BYN'
	| 'NIO'
	| 'LAK'
	| 'MDL'
	| 'STN'
	| 'CDF'
	| 'LBP'
	| 'MMK'
	| 'TMT'
	| 'WST'
	| 'BAM'

export type Lang =
	| 'ara'
	| 'eng'
	| 'fra'
	| 'fas'
	| 'ukr'
	| 'ind'
	| 'msa'
	| 'ces'
	| 'slk'
	| 'srp'
	| 'hmo'
	| 'tpi'
	| 'nau'
	| 'spa'
	| 'ron'
	| 'tvl'
	| 'zho'
	| 'sin'
	| 'tam'
	| 'afr'
	| 'nbl'
	| 'nso'
	| 'sot'
	| 'ssw'
	| 'tsn'
	| 'tso'
	| 'ven'
	| 'xho'
	| 'zul'
	| 'nfr'
	| 'swa'
	| 'dan'
	| 'fao'
	| 'smo'
	| 'tkl'
	| 'hin'
	| 'pih'
	| 'nep'
	| 'rus'
	| 'uzb'
	| 'fij'
	| 'hif'
	| 'mfe'
	| 'cnr'
	| 'gil'
	| 'por'
	| 'fil'
	| 'nor'
	| 'nld'
	| 'mri'
	| 'nzs'
	| 'lav'
	| 'tha'
	| 'arc'
	| 'ckb'
	| 'kin'
	| 'grn'
	| 'fin'
	| 'swe'
	| 'ita'
	| 'zdj'
	| 'mlt'
	| 'ell'
	| 'deu'
	| 'her'
	| 'hgm'
	| 'kwn'
	| 'loz'
	| 'ndo'
	| 'est'
	| 'ton'
	| 'urd'
	| 'glv'
	| 'aym'
	| 'que'
	| 'tur'
	| 'kir'
	| 'kon'
	| 'lin'
	| 'lat'
	| 'khm'
	| 'ber'
	| 'sqi'
	| 'hrv'
	| 'hye'
	| 'pau'
	| 'cat'
	| 'bjz'
	| 'hat'
	| 'sag'
	| 'de'
	| 'kal'
	| 'tgk'
	| 'som'
	| 'div'
	| 'crs'
	| 'pap'
	| 'mlg'
	| 'bwg'
	| 'kck'
	| 'khi'
	| 'ndc'
	| 'nde'
	| 'nya'
	| 'sna'
	| 'toi'
	| 'zib'
	| 'kor'
	| 'tir'
	| 'jam'
	| 'mon'
	| 'ltz'
	| 'mkd'
	| 'dzo'
	| 'kaz'
	| 'gsw'
	| 'roh'
	| 'cha'
	| 'nno'
	| 'nob'
	| 'smi'
	| 'bos'
	| 'prs'
	| 'pus'
	| 'tuk'
	| 'slv'
	| 'isl'
	| 'bul'
	| 'kat'
	| 'gle'
	| 'run'
	| 'cal'
	| 'lit'
	| 'nrf'
	| 'ben'
	| 'mey'
	| 'vie'
	| 'hun'
	| 'amh'
	| 'bis'
	| 'mah'
	| 'jpn'
	| 'aze'
	| 'heb'
	| 'rar'
	| 'tet'
	| 'pol'
	| 'bel'
	| 'pov'
	| 'lao'
	| 'niu'
	| 'lua'
	| 'mya'

export type Transl =
	| 'ara'
	| 'bre'
	| 'ces'
	| 'cym'
	| 'deu'
	| 'est'
	| 'fin'
	| 'fra'
	| 'hrv'
	| 'hun'
	| 'ita'
	| 'jpn'
	| 'kor'
	| 'nld'
	| 'per'
	| 'pol'
	| 'por'
	| 'rus'
	| 'slk'
	| 'spa'
	| 'srp'
	| 'swe'
	| 'tur'
	| 'urd'
	| 'zho'

export type DemonymNames = 'eng' | 'fra'